implement validation of plmxml targets vs. scene materials
Two validations are implemented. One general comparison in the global validation before
configuration(which can be omitted by the user). One at material update time which
is specific to the current configuration.
The first method should make sure that the material is actually assigned to
a geometry(acquired via SceneStructureRequest). While the config specific acquires
all scene materials. Potentially materials that are only available inside the look
library and not assigned to a geometry.